+++ shareaza/DlgSecureRule.cpp
@@ -113,7 +113,7 @@ BOOL CSecureRuleDlg::OnInitDialog()
 	m_nType		= m_pRule->m_nType;
 	m_sComment	= m_pRule->m_sComment;
 	m_nAction	= m_pRule->m_nAction;
-	m_nExpire	= min( m_pRule->m_nExpire, 2 );
+	m_nExpire	= min( m_pRule->m_nExpire, DWORD(2) );
 	m_nMatch	= 0;
 
 	if ( m_pRule->m_nType == CSecureRule::srAddress )
