+++ shareaza/PageSettingsBitTorrent.cpp
@@ -110,7 +110,7 @@ BOOL CBitTorrentSettingsPage::OnInitDialog()
 	m_wndClearPercentage.EnableWindow( m_bAutoClear );
 
 	DWORD nMaxTorrents = ( Settings.GetOutgoingBandwidth() / 2 ) + 2;
-	nMaxTorrents = min (10, nMaxTorrents);
+	nMaxTorrents = min (DWORD(10), nMaxTorrents);
 
 	m_wndClearPercentageSpin.SetRange( 100, 999 );
 
@@ -124,7 +124,7 @@ BOOL CBitTorrentSettingsPage::OnInitDialog()
 BOOL CBitTorrentSettingsPage::OnSetActive() 
 {
 	DWORD nMaxTorrents = ( Settings.GetOutgoingBandwidth() / 2 ) + 2;
-	nMaxTorrents = min (10, nMaxTorrents);
+	nMaxTorrents = min (DWORD(10), nMaxTorrents);
 
 	m_nDownloads	= min( m_nDownloads, (int)nMaxTorrents );
 	m_wndDownloadsSpin.SetRange( 0, (WORD)nMaxTorrents );
