+++ shareaza/WizardConnectionPage.cpp
@@ -309,9 +309,9 @@ LRESULT CWizardConnectionPage::OnWizardNext()
 	if ( theApp.m_bLimitedConnections ) 
 	{	//Window XP Service Pack 2
 		theApp.Message( MSG_ERROR, _T("Warning  - Windows XP Service Pack 2 detected. Performance may be reduced.") );
-		Settings.Downloads.ConnectThrottle		= max( Settings.Downloads.ConnectThrottle, 800 );
-		Settings.Connection.ConnectThrottle		= max( Settings.Connection.ConnectThrottle, 250 );
-		Settings.Gnutella.ConnectFactor			= min( Settings.Gnutella.ConnectFactor, 3 );
+		Settings.Downloads.ConnectThrottle		= max( Settings.Downloads.ConnectThrottle, DWORD(800) );
+		Settings.Connection.ConnectThrottle		= max( Settings.Connection.ConnectThrottle, DWORD(250) );
+		Settings.Gnutella.ConnectFactor			= min( Settings.Gnutella.ConnectFactor, DWORD(3) );
 		Settings.Gnutella2.NumHubs				= min( Settings.Gnutella2.NumHubs, 2 );
 		Settings.Gnutella1.EnableAlways			= FALSE;
 		Settings.Gnutella1.EnableToday			= FALSE;
